requirements.txt
flask==2.1.0
openai==1.3.0
python-dotenv==0.21.0
PyPDF2==3.0.1
python-docx==0.8.11
textract==1.6.0
flask-cors==3.0.10
gunicorn==20.1.0
